refactor: move instance-blocklist.ts to policy/ (Plan E)
